pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "boilerworks"
version = "0.1.0"
description = "Production-ready project templates — assembled in seconds."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
keywords = ["boilerplate", "templates", "scaffolding", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "click>=8.1",
    "questionary>=2.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "jinja2>=3.0",
    "gitpython>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
boilerworks = "boilerworks.cli:main"

[project.urls]
Homepage = "https://boilerworks.dev"
Repository = "https://github.com/ConflictHQ/boilerworks"
Issues = "https://github.com/ConflictHQ/boilerworks/issues"

[tool.hatch.build.targets.wheel]
packages = ["boilerworks"]

[tool.hatch.build.targets.sdist]
include = ["boilerworks/", "tests/"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "N"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=boilerworks --cov-report=term-missing --cov-fail-under=80"
filterwarnings = ["ignore::pytest.PytestCollectionWarning"]

[tool.coverage.run]
omit = ["tests/*"]
